% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbs.R
\name{as.vbslines}
\alias{as.vbslines}
\alias{as.vbslines.default}
\alias{as.vbslines.character}
\title{Coercion to \code{vbslines}}
\usage{
as.vbslines(x, ...)

\method{as.vbslines}{default}(x, ...)

\method{as.vbslines}{character}(x, ...)
}
\arguments{
\item{x}{any R object}

\item{...}{additional arguments passed to methods}
}
\description{
Coerce
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: default method --- only raises error

\item \code{character}: coerces character strings to class \ocde{vbslines}
}}

