% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{cache_metadata}
\alias{cache_metadata}
\alias{cache_metadata_file_name}
\alias{cache_metadata_file_path}
\alias{cache_metadata_read}
\alias{cache_metadata_write}
\alias{cache_metadata_append_or_replace}
\alias{cache_metadata_refresh}
\alias{cache_hash_random}
\alias{cache_hash}
\alias{cache_clean_hash}
\alias{cache_clean_all}
\title{Cache Metadata}
\usage{
cache_metadata_file_name()

cache_metadata_file_path()

cache_metadata_read()

cache_metadata_write(metadata)

cache_metadata_append_or_replace(hash, working.dir, input.file.path)

cache_metadata_refresh()

cache_hash_random()

cache_hash(df)

cache_clean_hash(hash)

cache_clean_all()
}
\arguments{
\item{metadata}{\verb{[data.frame]} (no default)

Metadata.}

\item{hash}{\verb{[character]} (no default)

Hash of an input dataset.}

\item{working.dir}{\verb{[character]} (no default)

Path to directory to include in metadata.}

\item{input.file.path}{\verb{[character]} (no default)

Path to input file (which must exist) to include in metadata.}

\item{df}{\verb{[data.frame]} (no default)
\code{data.frame} for which to compute hash.}
}
\description{
Tools to access and handle cache metadata. These are advanced /
under-the-hood tools which you almost certainly should not use for anything.
\verb{[iarccrgtools::cache_clean_hash]} and
\verb{[iarccrgtools::cache_clean_all]} may be the exceptions.
}
\section{Functions}{

\verb{[iarccrgtools::cache_metadata_file_path]} returns a string, the name of
to the cache metadata file.


\verb{[iarccrgtools::cache_metadata_file_path]} returns a string, the path
to the cache metadata.


\verb{[iarccrgtools::cache_metadata_read]} reads cache metadata from the path
given by
\verb{[iarccrgtools::cache_metadata_file_path]}. Returns a \code{data.table} with
zero rows if no file found.


\verb{[iarccrgtools::cache_metadata_write]} writes \code{metadata} to the path given by
\verb{[iarccrgtools::cache_metadata_file_path]}.


\verb{[iarccrgtools::cache_metadata_append_or_replace]} appends data to
the cache metadata, or replaces data that already exists for the
same input file path, if any.


\verb{[iarccrgtools::cache_metadata_refresh]} removes any non-existing
directories from the cache metadata.


\verb{[iarccrgtools::cache_hash_random]} outputs a random hash.


\verb{[iarccrgtools::cache_hash]} outputs a hash for the input \code{data.frame}.


\verb{[iarccrgtools::cache_clean_hash]} removes all cache dirs for the given hash.
Also cleans up the cache metadata.


\verb{[iarccrgtools::cache_clean_all]} removes all cache dirs.
Also cleans up the cache metadata.
}

