% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbs.R
\name{can_call_vbs}
\alias{can_call_vbs}
\title{Test .vbs Capability}
\usage{
can_call_vbs()
}
\description{
Simply returns TRUE/FALSE depending on whether you can execute
.vbs scripts. The reason may be due to your system not supporting them
or that you don't have sufficient permissions.
}
