% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_programs.R
\name{column_names}
\alias{column_names}
\alias{tool_colnameset_names}
\alias{tool_colnameset}
\alias{tool_column_classes}
\alias{tool_column_fwf_widths}
\alias{tool_column_infos}
\title{Column Names}
\usage{
tool_colnameset_names()

tool_colnameset(set.nm)

tool_column_classes(col.nms)

tool_column_fwf_widths(col.nms)

tool_column_infos(col.nms)
}
\arguments{
\item{set.nm}{string; name of a column set; only those returned by
\code{tool_colnameset_names} are allowed}

\item{col.nms}{character string vector; names of columns; must be subset of
items returned by \code{tool_colnameset("all")}}
}
\description{
Utilities for handling sets of required column names for each
IARC CRG Tools tool.
}
\section{Functions}{
\itemize{
\item \code{tool_colnameset_names}: each column name set has its own name. This function
returns all the names of the sets.

\item \code{tool_colnameset}: returns a set of column name given the name of the
set itself.

\item \code{tool_column_classes}: returns the required class for each column by
column name

\item \code{tool_column_fwf_widths}: returns an integer vector of default column widths
for fixed-width format

\item \code{tool_column_infos}: returns short plain English explanation of column
}}

