% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{column_specifications}
\alias{column_specifications}
\title{Program Definitions}
\format{A data.frame with character string columns
\itemize{
\item \code{column_name} name of column
\item \code{class} expected class of column when using functions in this package
\item \code{info} short plain English explanation of column purpose and contents
}

as well as a number of logical (TRUE/FALSE) columns. The names of these
logical columns follow the convention \code{"set_SETNAME"} where \code{SETNAME}
is one of the items returned by \code{\link{tools_program_colnameset_names}}.
Each such logical columns is \code{TRUE} when that column indicated in
\code{column_name} is included in that set of column names.}
\usage{
column_specifications
}
\description{
data.frame of specifications for columns used in various
functions
}
\seealso{
Other program_definition_data: \code{\link{program_guides}},
  \code{\link{program_output_files}}
}
\concept{program_definition_data}
\keyword{datasets}
