% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{create_example}
\alias{create_example}
\title{Example Datasets}
\usage{
create_example(...)
}
\arguments{
\item{...}{Unused, this function has been deprecated.}
}
\description{
DEPRECATED, use \verb{[iarccrgtools::tool_colnameset_example_dataset]}.
}
