% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executable.R
\name{exe_path}
\alias{exe_path}
\alias{set_path_to_exe}
\alias{get_path_to_exe}
\alias{guess_path_to_exe}
\title{IARC CRG Tools Executable Path}
\usage{
set_path_to_exe(path)

get_path_to_exe()

guess_path_to_exe()
}
\arguments{
\item{path}{path to IARC CRG Tools executable}
}
\description{
Set and get path to IARC CRG Tools executable.
}
\details{
The executable is attempted to be discovered automatically by
\code{\link{guess_path_to_exe}} at start-up of this package, so with
any luck you don't need to set it by hand.

Executable path will be stored for the duration of the ongoing R session.
If this package is somehow unloaded or reloaded, the setting will reset.
}
\section{Functions}{
\itemize{
\item \code{set_path_to_exe}: sets path to .exe

\item \code{get_path_to_exe}: gets currently saved path to .exe;
value will be \code{NA_character_} if no path has been set yet.

\item \code{guess_path_to_exe}: tries to guess where the executable is based on
typical installation directories
}}

