% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_width_format.R
\name{fwf}
\alias{fwf}
\alias{format_fwf}
\alias{format_fwf.integer}
\alias{format_fwf.numeric}
\alias{format_fwf.character}
\alias{format_fwf.Date}
\alias{write_fwf}
\alias{read_fwf}
\title{Fixed Width Format}
\usage{
format_fwf(x, width)

\method{format_fwf}{integer}(x, width)

\method{format_fwf}{numeric}(x, width)

\method{format_fwf}{character}(x, width)

\method{format_fwf}{Date}(x, width)

write_fwf(x, path, widths = NULL, ...)

read_fwf(path, ...)
}
\arguments{
\item{x}{\verb{[various]} (mandatory, no default)
\itemize{
\item \code{write_fwf}: \code{data.frame} to write to disk
\item \code{format_fwf}: an R object to format
}}

\item{width}{\verb{[integer]} (mandatory, no default)
width of field to be written}

\item{path}{\verb{[character]} (mandatory, no default)
\itemize{
\item \code{write_fwf}: path to write to
\item \code{read_fwf}: path to read from
}}

\item{widths}{\verb{[NULL, integer]} (optional, default \code{NULL})
\itemize{
\item \code{NULL}: the width of each field will be \code{max(nchar(x[[i]]))} for all
\verb{i in 1:ncol(x)}
\item \code{integer}: use these widths; must be of length \code{ncol(x)}
}}

\item{...}{additional arguments passed to
\itemize{
\item \code{write_fwf}: \link[data.table:fwrite]{data.table::fwrite}
\item \code{read_fwf}: \link[data.table:fread]{data.table::fread}
}}
}
\description{
Read and write fixed-width format files.
}
