% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_programs.R
\name{get_tools_settings_template}
\alias{get_tools_settings_template}
\title{IARC CRG Tools Settings Templates}
\usage{
get_tools_settings_template(dir.path = get_tool_work_dir(tool.name), tool.name)
}
\arguments{
\item{dir.path}{a character string; a path to a directory (that must exist)}

\item{tool.name}{\verb{[character]} (mandatory, no default)

name of tool to run; see output of \code{tool_clean_names()} for options}
}
\description{
Copy a settings file template
(see \code{\link{tools_settings_files}})
into a local directory. The settings may not be fully what they should
be for your use-case.
}
\details{
If a file with the same name as the one copied by this function already
exists in the directory, this function will ask whether to overwrite it
or not.
}
