% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{iarc_input_write}
\alias{iarc_input_write}
\alias{write_tools_data}
\title{Write File to Use as IARC CRG Tools Input}
\usage{
iarc_input_write(
  x,
  colnameset.name = tool_colnameset_names()[1],
  file = tempfile(fileext = ".txt", tmpdir = iarccrgtools::iarc_workdir_get()),
  overwrite = NULL,
  verbose = FALSE,
  fwrite_arg_list = NULL
)

write_tools_data(
  x,
  colnameset.name = tool_colnameset_names()[1],
  file = tempfile(fileext = ".txt", tmpdir = iarccrgtools::iarc_workdir_get()),
  overwrite = NULL,
  verbose = FALSE,
  fwrite_arg_list = NULL
)
}
\arguments{
\item{x}{a data.frame}

\item{colnameset.name}{one of the items returned by
\code{\link{tool_colnameset_names}}}

\item{file}{\verb{[character]} (mandatory, no default)

path where \code{x} will be written to; if the file already exists, this
function prompts you whether to overwrite or not; in non-interactive
use}

\item{overwrite}{\verb{[NULL, logical]} (optional, default \code{NULL})
\itemize{
\item \code{NULL}: if \code{file} already exists, user is prompted whether to overwrite;
in non-interactive mode (see \code{?interactive}) a pre-existing file causes
an error
\item \code{TRUE}: any pre-existing file is overwritten without prompting
\item \code{FALSE} any pre-existing file is overwritten causes an error
}}

\item{verbose}{\verb{[logical]} (mandatory, no default)

if \code{TRUE}, the function emits messages during it's run to let you know
what's happening}

\item{fwrite_arg_list}{\verb{[NULL, list]}

Passed to \link[fwf:fwf]{fwf::fwf_write} arg \code{fwrite_arg_list}. The following defaults are
internally:
\itemize{
\item \code{sep = ";"}
\item \code{dec = ","}
\item \code{quote = FALSE}
\item \code{row.names = FALSE}
\item \code{col.names = FALSE}
}}
}
\description{
A text file is written in a fixed field format.
}
\section{Functions}{
\itemize{
\item \code{write_tools_data()}: Deprecated --- use
\code{iarccrgtools::iarc_input_write}.

}}
