% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{iarc_result_read}
\alias{iarc_result_read}
\alias{read_tools_results}
\title{IARC CRG Tools Results}
\usage{
iarc_result_read(tool.name, hash, input.col.nms = NULL, verbose = TRUE)

read_tools_results(tool.name, hash, input.col.nms = NULL, verbose = TRUE)
}
\arguments{
\item{tool.name}{\verb{[character]} (mandatory, no default)

name of tool to run; see output of \code{tool_clean_names()} for options}

\item{hash}{\verb{[character]} (no default)
Hash of input dataset to read into R. See \verb{[iarccrgtools::cache_hash]}.}

\item{input.col.nms}{\code{NULL} (default) or a character string vector of column
names; when not \code{NULL}, allows setting column names on the tables that were
read. See Details.}

\item{verbose}{if TRUE, this function gives messages during its process}
}
\value{
A list of data.frames (where the read file was a table) and/or character
string vectors (where the read file was a non-table file such as a log file).
The names of the list correspond to the names of the files that were read.
}
\description{
Read IARC CRG Tools results into R.
}
\details{
When \code{input.col.nms} is a character string vector, it is attempted to be
used as the set of column names for any tables that are read by this
function. You should supply the same column names that were in the file
saved for use by IARC CRG Tools. If IARC CRG Tools adds a column in addition
to the ones you had in input data, that column will gain the name
\code{"tool_text"} automatically.
}
\section{Functions}{
\itemize{
\item \code{read_tools_results()}: Deprecated --- use
\code{iarccrgtools::iarc_result_read}.

}}
