% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_iarccrgtools.R
\docType{package}
\name{iarccrgtools}
\alias{iarccrgtools-package}
\alias{iarccrgtools}
\title{\code{iarccrgtools}}
\description{
Utilities for using IARC CRG Tools on Windows via R.

\href{https://github.com/WetRobot/iarccrgtools/actions/workflows/R-CMD-check.yaml}{\if{html}{\figure{https://github.com/WetRobot/iarccrgtools/actions/workflows/R-CMD-check.yaml/badge.svg}}}
}
\section{Recommended installation}{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{devtools::install_github(
  "FinnishCancerRegistry/iarccrgtools",
  ref = readline("enter latest tag on github: ")
)
}\if{html}{\out{</div>}}
}

\section{Example}{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{dir_path <- tempdir()
iarccrgtools::iarc_workdir_set(dir_path)

tool_name <- "check"
subset <- "mandatory"
iarc_df <- iarccrgtools::tool_colnameset_example_dataset(
  paste0(subset, "_", tool_name), n.rows = 10L
)
results <- iarccrgtools::interact_with_tool(
  iarc_df, tool.name = tool_name, clean = TRUE
)

result_df <- iarccrgtools::connect_tool_results_to_observations(
  record.ids = iarc_df[["record_id"]], tool.results = results
)
}\if{html}{\out{</div>}}
}

\section{News}{

\subsection{News for version 0.3.0}{
\subsection{iarccrgtools::get_tool_exe_dir_path}{

\code{iarccrgtools::get_tool_exe_dir_path} no longer usable.
}

\subsection{iarccrgtools::get_tool_work_dir}{

\code{iarccrgtools::get_tool_work_dir} no longer usable.
}

\subsection{iarccrgtools::get_tools_exe_path}{

\code{iarccrgtools::get_tools_exe_path} no longer usable ---
use `iarccrgtools::iarc_exe_path_get``.
}

\subsection{iarccrgtools::get_tools_install_dir_path}{

\code{iarccrgtools::get_tools_install_dir_path} no longer usable.
}

\subsection{iarccrgtools::get_tools_work_dir}{

\code{iarccrgtools::get_tools_work_dir} no longer usable ---
use `iarccrgtools::iarc_workdir_get``.
}

\subsection{iarccrgtools::guess_tools_exe_dir_path}{

\code{iarccrgtools::guess_tools_exe_dir_path} no longer usable.
}

\subsection{iarccrgtools::guess_tools_exe_path}{

\code{iarccrgtools::guess_tools_exe_path} no longer usable ---
use `iarccrgtools::iarc_exe_path_guess``.
}

\subsection{iarccrgtools::iarc_exe_dir_path_guess}{

\code{iarccrgtools::iarc_exe_dir_path_guess} no longer usable.
}

\subsection{iarccrgtools::interact_with_tool}{

\code{iarccrgtools::interact_with_tool} now accepts \code{icdo3_topography} value with
up to five characters.

\code{iarccrgtools::interact_with_tool} now no longer requires \code{subject_id},
\code{record_id} to be of class integer --- it can be any class. Remember that
the class of each column in results read into R is ultimately decided by
\code{data.table::fread}.
}

\subsection{iarccrgtools::set_tools_exe_path}{

\code{iarccrgtools::set_tools_exe_path} no longer usable ---
use `iarccrgtools::iarc_exe_path_set``.
}

\subsection{iarccrgtools::set_tools_work_dir}{

\code{iarccrgtools::set_tools_work_dir} no longer usable ---
use `iarccrgtools::iarc_workdir_set``.
}

}

\subsection{News for version 0.2.30}{
\subsection{iarccrgtools::format_fwf}{

\code{iarccrgtools::format_fwf} permanently removed.
}

\subsection{iarccrgtools::read_fwf}{

\code{iarccrgtools::read_fwf} permanently removed. Use
\code{fwf::fwf_read}.
}

\subsection{iarccrgtools::write_fwf}{

\code{iarccrgtools::write_fwf} permanently removed. Use
\code{fwf::fwf_write}.
}

}

\subsection{News for version 0.2.29}{
\subsection{iarccrgtools::create_example}{

\code{iarccrgtools::create_example} permanently removed. Use
\code{iarccrgtools::tool_colnameset_example_dataset}.
}

}

\subsection{News for version 0.2.28}{
\subsection{iarccrgtools::interact_with_tool}{

If the virtual directory does not exist, it is now attempted to be
created. See ?iarccrgtools::interact_with_tool for information about
what the virtual directory is.
}

\subsection{iarccrgtools::interact_with_tool}{

Document what steps \code{iarccrgtools::interact_with_tool} performs, in
particular caching and parameter/setting files.
}

}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FinnishCancerRegistry/iarccrgtools}
  \item Report bugs at \url{https://github.com/FinnishCancerRegistry/iarccrgtools/issues}
}

}
