% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interface.R
\name{interface_with_tool}
\alias{interface_with_tool}
\alias{automate_tool}
\alias{interact_with_tool}
\title{IARC CRG Tools R Interface}
\usage{
interface_with_tool(
  data,
  tool.name,
  how = c("interactively", "automatically")[1],
  clean = FALSE,
  verbose = FALSE
)

automate_tool(data, tool.name, clean = FALSE, verbose = FALSE)

interact_with_tool(data, tool.name, clean = FALSE, verbose = FALSE)
}
\arguments{
\item{data}{a \code{data.frame}; required columns depend on
the IARC CRG Tools tool used}

\item{tool.name}{\verb{[character]} (mandatory, no default)

name of tool to run; see output of \code{tool_clean_names()} for options}

\item{how}{\verb{[character]} (mandatory, default \code{"interactively"})
\itemize{
\item \code{"interactively"}: the user must open IARC CRG Tools manually and interact
with it according to instructions, but the dataset is written to disk
and read back into R by this function
\item \code{"automatically"}: like \code{"interactively"}, but the appropriate tool
is attempted to be called without any user interaction
}}

\item{clean}{\verb{[logical]} (optional, default \code{FALSE})
\itemize{
\item \code{TRUE}: all input and output files for IARC CRG Tools will be removed
from disk after the results have been read into R
\item \code{FALSE}: all files are let to be in peace
}}

\item{verbose}{\verb{[logical]} (optional, default \code{FALSE})
\itemize{
\item \code{TRUE}: emit messages which are not needed for correct usage but which
tell at what stage the is process and can be helpful in debugging
\item \code{FALSE}: only necessary messages are emitted
}}
}
\description{
Open IARC CRG Tools and simulate keystrokes to run the tool from start
to finish.
}
\details{
See \code{\link{interact_with_tool}} for the manual but more foolproof
method.

This function requires that you are able to execute .vbs scripts.
IARC CRG Tools is opened and keystrokes are sent using such scripts.
You can test this using \code{\link{can_call_vbs}}.

Before using this function for the first time you need to run
\code{\link{interact_with_tool}} to build a settings file for future
use into the working directory set using \code{\link{set_tools_root_dir}}.
This must be done for each tool separately. Read more about the
settings files here: \code{\link{tools_settings_files}}.

After the settings file is in place for the intended tool, from then on
you can run this function with the same working directory set and the
tool runs from start to finish automatically. The data is first
written into the set working directory, then IARC CRG Tools is run, and
the resulting files are read into R.

This function should not be considered fool-proof. Currently this function
assumes that IARC CRG Tools has finished its computations when the
file size of the tool output has not increased in 30 seconds. In edge
cases this may be incorrect. Additionally, no error-recovery logic
has been written for this function in case IARC CRG Tools is interrupted
or raises an error otherwise. Hence do not rely on this function for critical
processes. However the approach used here appears to work
fine in practice when IARC CRG Tools is not interrupted.
}
\seealso{
\code{\link{interact_with_tool}}
}
