% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{program_guides}
\alias{program_guides}
\title{Program Definitions}
\format{A data.frame with these character string columns:
\itemize{
\item \code{program_name}: name of the program
\item \code{command}: computer-readable specification of keystrokes and certain special
commands (see below) that are executed in the given order to run an
IARC CRG Tools program from start to finish; mainly relevant for
\code{\link{use_tools_automatically}}
\item \code{instruction}: plain English explanation of each step; these are shown
when you call \code{\link{use_tools_interactively}}, so they should
tell you quite clearly what you need to do.
}}
\usage{
program_guides
}
\description{
data.frame of commands for assisted or
automated use of IARC CRG Tools.
}
\seealso{
Other program_definition_data: \code{\link{program_output_files}}
}
\concept{program_definition_data}
\keyword{datasets}
