% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{program_output_files}
\alias{program_output_files}
\title{Program Definitions}
\format{A data.frame with these columns:
\itemize{
\item \code{program_name}: character string column; name of the program
\item \code{file_name_suffix}: character string column; suffix pasted to each output
file name; in other words the output files are assumed to have these
suffixes
\item \code{is_table}: logical column; \code{TRUE} if the output file is a table,
\code{FALSE} if it is non-tabular text (such as a log file)
}}
\usage{
program_output_files
}
\description{
data.frame specifying output files for each
IARC CRG Tools program.
}
\seealso{
Other program_definition_data: \code{\link{column_specifications}},
  \code{\link{program_guides}}
}
\concept{program_definition_data}
\keyword{datasets}
