% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_tools_results}
\alias{read_tools_results}
\title{IARC CRG Tools Results}
\usage{
read_tools_results(program.name)
}
\arguments{
\item{program.name}{character string; name of program to run}
}
\description{
Read IARC CRG Tools results into R.
}
