% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_programs.R
\name{tool_names}
\alias{tool_names}
\title{IARC CRG Tools Program Names}
\usage{
tool_names()
}
\value{
A character string vector.
}
\description{
This function simply returns a haracter string vector of IARC CRG Tools
tool names supported by this R package.
}
