% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{tool_output_files}
\alias{tool_output_files}
\title{Program Definitions}
\format{
A data.frame with these columns:
\itemize{
\item \code{tool_name}: character string column; name of the program
\item \code{file_name_suffix}: character string column; suffix pasted to each output
file name; in other words the output files are assumed to have these
suffixes
\item \code{is_table}: logical column; \code{TRUE} if the output file is a table,
\code{FALSE} if it is non-tabular text (such as a log file)
}
}
\usage{
tool_output_files
}
\description{
data.frame specifying output files for each
IARC CRG Tools tool.
}
\seealso{
Other tool_definition_data: 
\code{\link{column_specifications}},
\code{\link{tools}}
}
\concept{tool_definition_data}
\keyword{datasets}
