% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_settings.R
\name{tool_settings}
\alias{tool_settings}
\alias{tool_settings_source_dir_path}
\alias{tool_settings_availability}
\alias{tool_settings_are_available}
\alias{tool_settings_copy}
\title{IARC CRG Tools Settings}
\usage{
tool_settings_source_dir_path()

tool_settings_availability()

tool_settings_are_available(colnameset.name)

tool_settings_copy(tgt.dir.path, colnameset.name, verbose = TRUE)
}
\arguments{
\item{colnameset.name}{one of the items returned by
\code{\link{tool_colnameset_names}}}

\item{tgt.dir.path}{\verb{[character]} (no default)

Path to an existing directory.}

\item{verbose}{\verb{[logical]} (default \code{TRUE})

If \code{TRUE}, a message is emitted concerning what was copied where, if
anything.}

\item{dir.path}{a character string; a path to a directory (that must exist)}
}
\description{
Tools to handle settings files included in this package.
You almost certainly do not need to use any of this functions yourself.
Instead, other functions in this package use them.
}
\section{Functions}{

\verb{[iarccrgtools::tool_settings_source_dir_path]} returns a string, the
location of the settings files inside this R package.


\verb{[iarccrgtools::tool_settings_availability]} returns a \code{data.table}
with columns \code{file_name}, \code{file_path}, \code{colnameset_name},
\code{tool_name}. This dataset describes for which column name sets a settings
file is available in this R package.


\verb{[iarccrgtools::tool_settings_are_available]} returns a \code{TRUE}
if a settings file is available for \code{colnameset.name} within this R package.


\verb{[iarccrgtools::tool_settings_copy]} copies the settings file(s) from within
this R package to \code{tgt.dir.path}.
}

\examples{

# iarccrgtools::tool_settings_source_dir_path
stopifnot(
  dir.exists(iarccrgtools::tool_settings_source_dir_path())
)


# iarccrgtools::tool_settings_availability
stopifnot(
  is.data.frame(iarccrgtools::tool_settings_availability())
)


# iarccrgtools::tool_settings_are_available
stopifnot(
  iarccrgtools::tool_settings_are_available("mandatory_check"),
  iarccrgtools::tool_settings_are_available("all_multiple_primary"),
  iarccrgtools::tool_settings_are_available("mandatory_icdo3_to_icd10")
)

}
