% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_specific_metadata_tools.R
\name{tool_specific_metadata}
\alias{tool_specific_metadata}
\alias{tool_menu_name}
\alias{tool_menu_item_name}
\alias{tool_names}
\alias{tool_colnameset_name_to_tool_name}
\alias{tool_colnameset_names}
\alias{tool_colnameset}
\alias{tool_colnameset_example_dataset}
\alias{tool_column_classes}
\alias{tool_column_fwf_widths}
\alias{tool_column_explanation}
\title{Tool-specific Metadata}
\usage{
tool_menu_name(tool.name)

tool_menu_item_name(tool.name)

tool_names()

tool_colnameset_name_to_tool_name(colnameset.name)

tool_colnameset_names()

tool_colnameset(colnameset.name)

tool_colnameset_example_dataset(colnameset.name, n.rows = 10L)

tool_column_classes(col.nms)

tool_column_fwf_widths(col.nms)

tool_column_explanation(col.nms)
}
\arguments{
\item{tool.name}{\verb{[character]} (mandatory, no default)

name of tool to run; see output of \code{tool_clean_names()} for options}

\item{colnameset.name}{one of the items returned by
\code{\link{tool_colnameset_names}}}

\item{n.rows}{\verb{[integer]} (default \code{10L})

How many times should the example row (only one is defined) be repeated?
Each resulting row will have its own \code{subject_id} and \code{record_id}.}

\item{col.nms}{\verb{[character]} (no default)

Names of columns.}
}
\description{
Tools to access tool-specific metadata. You will almost certainly not need
to use any of these functions.
}
\section{Functions}{

\verb{[iarccrgtools::tool_menu_name]} gives the name of the menu in IARC CRG Tools
for the given \code{tool.name}.


\verb{[iarccrgtools::tool_menu_item_name]} gives the name of the item in the
appropriate menu in IARC CRG Tools for the given \code{tool.name}.


\verb{[iarccrgtools::tool_names]} gives the names of all tools.


\verb{[iarccrgtools::tool_colnameset_name_to_tool_name]} gets the name of the tool
from the column name set name.


\verb{[iarccrgtools::tool_colnameset_names]} returns all names of column name
sets.


\verb{[iarccrgtools::tool_colnameset]} returns all column names for the given
column name set name.


\verb{[iarccrgtools::tool_example_dataset]} returns a \code{data.table},
a fake example of what the data should look like for a given
\code{colnameset.name}.


\verb{[iarccrgtools::tool_column_classes]} returns the appropriate class for each
supplied column name.


\verb{[iarccrgtools::tool_column_fwf_widths]} returns the column width for each
given column name (width in the text file which IARC CRG Tools uses).


\verb{[iarccrgtools::tool_column_explanation]} gives as short explanation for each
given column name.
}

\examples{

# iarccrgtools::tool_menu_name
stopifnot(
  iarccrgtools::tool_menu_name("check") == "Tools"
)

# iarccrgtools::tool_menu_item_name
stopifnot(
  iarccrgtools::tool_menu_item_name("check") == "IARC/IACR Check"
)

# iarccrgtools::tool_names
stopifnot(
  c("check", "icdo3_to_icd10") \%in\% iarccrgtools::tool_names()
)

# iarccrgtools::tool_colnameset_name_to_tool_name
stopifnot(
  iarccrgtools::tool_colnameset_name_to_tool_name("mandatory_check") ==
    "check"
)

# iarccrgtools::tool_colnameset_names
stopifnot(
  "all_check" \%in\% iarccrgtools::tool_colnameset_names()
)

# iarccrgtools::tool_colnameset
stopifnot(
  "sex" \%in\% iarccrgtools::tool_colnameset("all_check")
)

# iarccrgtools::tool_colnameset_example_dataset
ac <- iarccrgtools::tool_colnameset_example_dataset("all_check", 20L)
stopifnot(
  "sex" \%in\% names(ac),
  !duplicated(ac[["subject_id"]]),
  vapply(names(ac), function(ac_col_nm) {
    inherits(ac[[ac_col_nm]], iarccrgtools::tool_column_classes(ac_col_nm))
  }, logical(1L))
)

# iarccrgtools::tool_column_classes
stopifnot(
  iarccrgtools::tool_column_classes("sex") == "integer"
)

# iarccrgtools::tool_column_fwf_widths
stopifnot(
  iarccrgtools::tool_column_fwf_widths("sex") == 1
)

# iarccrgtools::tool_column_explanation
stopifnot(
  iarccrgtools::tool_column_explanation("dg_date") == "Date of diagnosis"
)
}
