% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{tools}
\alias{tools}
\title{Program Definitions}
\format{
A data.frame with these character string columns:
\itemize{
\item \code{clean_name}: name of tool as understood by the functions in this
package
\item \code{real_name}: name of tool in IARC CRG Tools menu
\item \code{executable_name}: name of executable for corresponding tool
\item \code{window_name}: name of executable's window in Windows
}
}
\usage{
tools
}
\description{
Program definitions dataset.
}
\seealso{
Other tool_definition_data: 
\code{\link{column_specifications}},
\code{\link{tool_output_files}}
}
\concept{tool_definition_data}
\keyword{datasets}
