% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_programs.R
\name{tools_program_names}
\alias{tools_program_names}
\title{IARC CRG Tools Program Names}
\usage{
tools_program_names()
}
\value{
A character string vector.
}
\description{
This function simply returns a haracter string vector of IARC CRG Tools
program names supported by this R package.
}
