% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interface.R
\name{use_tools_automatically}
\alias{use_tools_automatically}
\title{IARC CRG Tools R Interface}
\usage{
use_tools_automatically(tools.data, program.name)
}
\arguments{
\item{tools.data}{a \code{data.frame}; required columns depend on
the IARC CRG Tools program used}

\item{program.name}{character string; name of program to run}
}
\description{
Open IARC CRG Tools and simulate keystrokes to run the program from start
to finish.
}
\details{
See \code{\link{use_tools_interactively}} for the manual but more foolproof
method.

This function requires that you are able to execute .vbs scripts.
IARC CRG Tools is opened and keystrokes are sent using such scripts.
You can test this using \code{\link{can_call_vbs}}.

Before using this function for the first time you need to run
\code{\link{use_tools_interactively}} to build a settings file for future
use into the working directory set using \code{\link{set_tools_working_dir}}.
This must be done for each program separately. Read more about the
settings files here: \code{\link{tools_settings_files}}.

After the settings file is in place for the intended program, from then on
you can run this function with the same working directory set and the
program runs from start to finish automatically. The data is first
written into the set working directory, then IARC CRG Tools is run, and
the resulting files are read into R.

This function should not be considered fool-proof. Currently this function
assumes that IARC CRG Tools has finished its computations when the
file size of the program output has not increased in 30 seconds. In edge
cases this may be incorrect. Additionally, no error-recovery logic
has been written for this function in case IARC CRG Tools is interrupted
or raises an error otherwise. Hence do not rely on this function for critical
processes. However the approach used here appears to work
fine in practice when IARC CRG Tools is not interrupted.
}
\seealso{
\code{\link{use_tools_interactively}}
}
