% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_main_exe_tools.R
\name{exe_path}
\alias{exe_path}
\alias{iarc_exe_path_set}
\alias{set_tools_exe_path}
\alias{iarc_exe_path_get}
\alias{get_tools_exe_path}
\alias{iarc_exe_path_guess}
\alias{guess_tools_exe_path}
\alias{iarc_exe_dir_path_guess}
\alias{guess_tools_exe_dir_path}
\alias{get_tools_install_dir_path}
\alias{get_tool_exe_dir_path}
\title{IARC CRG Tools Executable Path}
\usage{
iarc_exe_path_set(path)

set_tools_exe_path(path)

iarc_exe_path_get()

get_tools_exe_path()

iarc_exe_path_guess()

guess_tools_exe_path()

iarc_exe_dir_path_guess()

guess_tools_exe_dir_path()

get_tools_install_dir_path()

get_tool_exe_dir_path()
}
\arguments{
\item{path}{\verb{[character]} (no defaukt)

Path to IARC CRG Tools main executable.}
}
\description{
Set and get path to IARC CRG Tools executable.
}
\details{
The executable is attempted to be discovered automatically by
\link{guess_tools_exe_path} at start-up of this package, so with
any luck you don't need to set it by hand.

Executable path will be stored for the duration of the ongoing R session.
If this package is somehow unloaded or reloaded, the setting will reset.
}
\section{Functions}{
\itemize{
\item \code{iarc_exe_path_set()}: Sets path to IARC CRG Tools main executable.

\item \code{set_tools_exe_path()}: Deprecated.

\item \code{iarc_exe_path_get()}: Get currently set path to IARC CRG Tools main
executable.

\item \code{get_tools_exe_path()}: Deprecated.

\item \code{iarc_exe_path_guess()}: Guesses location of IARC CRG Tools main executable
based on usual installation locations. This will work if you have
Installed IARC CRG Tools in either \verb{C:/Program Files (x86)/IARCcrgTools"}
or \code{"C:/Program Files/IARCcrgTools"}.

\item \code{guess_tools_exe_path()}: Deprecated.

\item \code{iarc_exe_dir_path_guess()}: Deprecated.

\item \code{guess_tools_exe_dir_path()}: Deprecated.

\item \code{get_tools_install_dir_path()}: Deprecated.

\item \code{get_tool_exe_dir_path()}: Deprecated.

}}
