% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_executable.R
\name{exe_path}
\alias{exe_path}
\alias{set_tools_exe_path}
\alias{get_tools_exe_path}
\alias{guess_tools_exe_dir_path}
\alias{get_tools_install_dir_path}
\alias{get_tool_exe_dir_path}
\alias{guess_tools_exe_path}
\title{IARC CRG Tools Executable Path}
\usage{
set_tools_exe_path(path)

get_tools_exe_path()

guess_tools_exe_dir_path()

get_tools_install_dir_path()

get_tool_exe_dir_path()

guess_tools_exe_path()
}
\arguments{
\item{path}{path to IARC CRG Tools executable}
}
\description{
Set and get path to IARC CRG Tools executable.
}
\details{
The executable is attempted to be discovered automatically by
\link{guess_tools_exe_path} at start-up of this package, so with
any luck you don't need to set it by hand.

Executable path will be stored for the duration of the ongoing R session.
If this package is somehow unloaded or reloaded, the setting will reset.
}
\section{Functions}{
\itemize{
\item \code{set_tools_exe_path()}: sets path to .exe

\item \code{get_tools_exe_path()}: gets currently saved path to .exe;
value will be \code{NA_character_} if no path has been set yet.

\item \code{guess_tools_exe_dir_path()}: tries to guess where the executable is

\item \code{get_tools_install_dir_path()}: gets currently saved path to directory where
IARC CRG Tools is installed

\item \code{get_tool_exe_dir_path()}: subdir of \code{iarccrgtools::get_tools_install_dir_path()} output where
tool-specfic executables live

\item \code{guess_tools_exe_path()}: tries to guess where the main executable is based on
typical installation directories

}}
