% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_iarccrgtools.R
\docType{package}
\name{iarccrgtools}
\alias{iarccrgtools}
\title{Package \code{iarccrgtools}}
\description{
iarccrgtools

Utilities for using IARC CRG Tools on Windows via R.
}
\section{News}{

\subsection{News for version 0.2.30}{
\subsection{Object iarccrgtools::format_fwf news:}{

\code{iarccrgtools::format_fwf} permanently removed.
}

\subsection{Object iarccrgtools::read_fwf news:}{

\code{iarccrgtools::read_fwf} permanently removed. Use
\code{fwf::fwf_read}.
}

\subsection{Object iarccrgtools::write_fwf news:}{

\code{iarccrgtools::write_fwf} permanently removed. Use
\code{fwf::fwf_write}.
}

}

\subsection{News for version 0.2.29}{
\subsection{Object iarccrgtools::create_example news:}{

\code{iarccrgtools::create_example} permanently removed. Use
\code{iarccrgtools::tool_colnameset_example_dataset}.
}

}

\subsection{News for version 0.2.28}{
\subsection{Object iarccrgtools::interact_with_tool news:}{

If the virtual directory does not exist, it is now attempted to be
created. See ?iarccrgtools::interact_with_tool for information about
what the virtual directory is.
}

\subsection{Object iarccrgtools::interact_with_tool news:}{

Document what steps \code{iarccrgtools::interact_with_tool} performs, in
particular caching and parameter/setting files.
}

}
}

