% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_iarccrgtools.R
\docType{package}
\name{iarccrgtools}
\alias{iarccrgtools}
\title{Package \code{iarccrgtools}}
\description{
iarccrgtools

Utilities for using IARC CRG Tools on Windows via R.
}
\section{News}{

\subsection{News for version 0.3.0}{
\subsection{Object iarccrgtools::get_tool_exe_dir_path news:}{

\code{iarccrgtools::get_tool_exe_dir_path} no longer usable.
}

\subsection{Object iarccrgtools::get_tool_work_dir news:}{

\code{iarccrgtools::get_tool_work_dir} no longer usable.
}

\subsection{Object iarccrgtools::get_tools_exe_path news:}{

\code{iarccrgtools::get_tools_exe_path} no longer usable ---
use `iarccrgtools::iarc_exe_path_get``.
}

\subsection{Object iarccrgtools::get_tools_install_dir_path news:}{

\code{iarccrgtools::get_tools_install_dir_path} no longer usable.
}

\subsection{Object iarccrgtools::get_tools_work_dir news:}{

\code{iarccrgtools::get_tools_work_dir} no longer usable ---
use `iarccrgtools::iarc_workdir_get``.
}

\subsection{Object iarccrgtools::guess_tools_exe_dir_path news:}{

\code{iarccrgtools::guess_tools_exe_dir_path} no longer usable.
}

\subsection{Object iarccrgtools::guess_tools_exe_path news:}{

\code{iarccrgtools::guess_tools_exe_path} no longer usable ---
use `iarccrgtools::iarc_exe_path_guess``.
}

\subsection{Object iarccrgtools::iarc_exe_dir_path_guess news:}{

\code{iarccrgtools::iarc_exe_dir_path_guess} no longer usable.
}

\subsection{Object iarccrgtools::interact_with_tool news:}{

\code{iarccrgtools::interact_with_tool} now accepts \code{icdo3_topography} value with
up to five characters.

\code{iarccrgtools::interact_with_tool} now no longer requires \code{subject_id},
\code{record_id} to be of class integer --- it can be any class. Remember that
the class of each column in results read into R is ultimately decided by
\code{data.table::fread}.
}

\subsection{Object iarccrgtools::set_tools_exe_path news:}{

\code{iarccrgtools::set_tools_exe_path} no longer usable ---
use `iarccrgtools::iarc_exe_path_set``.
}

\subsection{Object iarccrgtools::set_tools_work_dir news:}{

\code{iarccrgtools::set_tools_work_dir} no longer usable ---
use `iarccrgtools::iarc_workdir_set``.
}

}

\subsection{News for version 0.2.30}{
\subsection{Object iarccrgtools::format_fwf news:}{

\code{iarccrgtools::format_fwf} permanently removed.
}

\subsection{Object iarccrgtools::read_fwf news:}{

\code{iarccrgtools::read_fwf} permanently removed. Use
\code{fwf::fwf_read}.
}

\subsection{Object iarccrgtools::write_fwf news:}{

\code{iarccrgtools::write_fwf} permanently removed. Use
\code{fwf::fwf_write}.
}

}

\subsection{News for version 0.2.29}{
\subsection{Object iarccrgtools::create_example news:}{

\code{iarccrgtools::create_example} permanently removed. Use
\code{iarccrgtools::tool_colnameset_example_dataset}.
}

}

\subsection{News for version 0.2.28}{
\subsection{Object iarccrgtools::interact_with_tool news:}{

If the virtual directory does not exist, it is now attempted to be
created. See ?iarccrgtools::interact_with_tool for information about
what the virtual directory is.
}

\subsection{Object iarccrgtools::interact_with_tool news:}{

Document what steps \code{iarccrgtools::interact_with_tool} performs, in
particular caching and parameter/setting files.
}

}
}

