% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interface.R
\name{interact_with_tool}
\alias{interact_with_tool}
\alias{connect_tool_results_to_observations}
\title{IARC CRG Tools R Interface}
\usage{
interact_with_tool(data, tool.name, clean = FALSE, verbose = FALSE)

connect_tool_results_to_observations(record.ids, tool.results)
}
\arguments{
\item{data}{a \code{data.frame}; required columns depend on
the IARC CRG Tools tool used}

\item{tool.name}{\verb{[character]} (mandatory, no default)

name of tool to run; see output of \code{tool_clean_names()} for options}

\item{clean}{\verb{[logical]} (optional, default \code{FALSE})
\itemize{
\item \code{TRUE}: all input and output files for IARC CRG Tools will be removed
from disk after the results have been read into R
\item \code{FALSE}: all files are left to be in peace
}}

\item{verbose}{if \code{TRUE}, this functions is chatty and gives messages
durings its run; if \code{FALSE}, you will only see necessary messages}

\item{record.ids}{\verb{[integer]} (mandatory, no default)

IDs of records for which to retrieve any record-specific results from
\code{tool.results}}

\item{tool.results}{\verb{[list]} (mandatory, no default)

list of tables and/or log texts as output by one of the interface functions
to IARC CRG Tools (e.g. \link{interact_with_tool})}
}
\description{
Write data for IARC CRG Tools and read it back into R after the run
has finished.
}
\details{
\itemize{
\item \code{connect_tool_results_to_observations} returns a \code{data.table} with
\code{length(record.ids)} rows; it has column
\code{record_id} and additional columns depending on results in
\code{tool.results}; this function goes through each object in \code{tool.results}
and if an object is a \code{data.table} with columns \code{record_id} and \code{tool_text},
each record appearing in that \code{data.table} is marked in the output
\code{data.table} in a logical column (e.g. \code{ in_multiple_primary_input.exl})
and any text in \code{tool_text} is collected into a separate column
(e.g. \code{multiple_primary_input.exl}); therefore the columns in the output of
\code{connect_tool_results_to_observations} vary by tool used.
}
}
