% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\name{tools_settings_files}
\alias{tools_settings_files}
\title{IARC CRG Tools Settings}
\description{
IARC CRG Tools can use settings defined on an earlier run of one of its
tools. This page documents the relationship of these settings files
with this R package.
}
\section{Manual use}{


This mainly pertains to using \code{\link{interact_with_tool}}.
When using IARC CRG Tools manually, settings files are not mandatory.
They just make your life easier. Inspect the help pages of IARC CRG Tools
for more information.
}

\section{Automatic use}{


This pertains to using \code{automate_tool}.
You need to have a pre-defined settings file for the IARC CRG Tools program
you want to use before using \code{automate_tool}.

Each tool must have its own settings file. The file extension will be
either .dfi or .frm. The name of the file must be the name of the program.
E.g.
\code{"iarc_check.dfi"}. The settings files must be stored in the
working directory set by \code{\link{set_tools_work_dir}}.

This R package has pre-defined "sensible defaults" for certain tools,
which you can fetch into a specific folder using
\code{\link{tool_settings_copy}}.
}

