% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interface.R
\name{use_tools_interactively}
\alias{use_tools_interactively}
\title{IARC CRG Tools R Interface}
\usage{
use_tools_interactively(tools.data, program.name)
}
\arguments{
\item{tools.data}{a \code{data.frame}; required columns depend on
the IARC CRG Tools program used}

\item{program.name}{character string; name of program to run}
}
\description{
Simplifies using IARC CRG Tools by preparing the data in a suitable
format and reading the data into R when appropriate. However, you still need
to use IARC CRG Tools manually.
}
\details{
See \code{\link{use_tools_automatically}} for the automatic but
less foolproof method.

This function saves the supplied data in a format useful for IARC CRG Tools
into the working directory set via \code{\link{set_tools_working_dir}},
prompts you to use the intended IARC CRG Tools program manually,
and reads the data back into R once you give the OK that IARC CRG Tools
has finished its computations.
}
