% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{work_dir}
\alias{work_dir}
\alias{set_tools_working_dir}
\alias{get_tools_working_dir}
\title{IARC CRG Tools Working Directory}
\usage{
set_tools_working_dir(dir)

get_tools_working_dir()
}
\arguments{
\item{dir}{string; path to a directory}
}
\description{
Get and set directory where input and output files of IARC CRG Tools
are stored.
}
\section{Functions}{
\itemize{
\item \code{set_tools_working_dir}: sets working directory

\item \code{get_tools_working_dir}: gets current working directory as string
}}

