% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_tools_data}
\alias{write_tools_data}
\title{Write File to Use as IARC CRG Tools Input}
\usage{
write_tools_data(
  x,
  colnameset.nm = tool_colnameset_names()[1],
  file = tempfile(fileext = ".txt", tmpdir = get_tools_root_dir()),
  ...
)
}
\arguments{
\item{x}{a data.frame}

\item{colnameset.nm}{one of the items returned by
\code{\link{tool_colnameset_names}}}

\item{file}{string; where file will be saved to}

\item{...}{arguments passed to \link{write_fwf};
e.g. try \code{nThread = x} where \code{x} is a desired number of cores to
use when writing}
}
\description{
A text file is written in a fixed field format.
}
